fix write_fmt related bug
this was a strange rabbithole. the aarch64-unknown-none (hardfloat)
implementation of `memcpy` uses a 128-bit arm neon register exactly
once, but the kernel runs with the fpu disabled or something, so that
traps.

switching to the aarch64-unknown-none-softfloat target makes
llvm/rustc not use neon, but it was linking Very Wrong. as in
producing an empty binary. it turned out that i had written a linker
script but never bothered to pass it to rustc, and the hardfloat
target was working anyway by chance. oops.